[ut] Add AutoTradingSystem
