+++ TradingSystem-CTeam/TradingSystem-CTeam.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="nemo_api.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="auto_trading_system.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
