+++ TradingSystem-CTeam/main.cpp
@@ -2,16 +2,9 @@
 #include <string>
 #include "gmock/gmock.h"
 
-using namespace testing;
+#include "auto_trading_system.cpp"
 
-//Interface Class
-class StockBrokerDriverInterface {
-public:
-    virtual void login(const std::string& id, const std::string& password) = 0;
-    virtual void buy(const std::string& stockCode, int count, int price) = 0;
-    virtual void sell(const std::string& stockCode, int count, int price) = 0;
-    virtual int currentPrice(const std::string& stockCode) = 0;
-};
+using namespace testing;
 
 //MockDirver
 class MockDriver : public StockBrokerDriverInterface {
